refactor(game-list): clean up GameListItemElement components
