+++ resources/js/features/game-list/components/GameListItems/GameListItemElement/GameListItemContent/index.ts
@@ -1,0 +1,1 @@
+export * from './GameListItemContent';
