+++ resources/js/features/game-list/components/GameListItems/useGameBacklogState.ts
@@ -8,14 +8,12 @@ interface UseGameBacklogStateProps {
   game: App.Platform.Data.Game;
   isInitiallyInBacklog: boolean;
 
-  shouldShowToasts?: boolean;
   shouldUpdateOptimistically?: boolean;
 }
 
 export function useGameBacklogState({
   game,
   isInitiallyInBacklog,
-  shouldShowToasts = true,
   shouldUpdateOptimistically = true,
 }: UseGameBacklogStateProps) {
   const { auth } = usePageProps();
@@ -40,64 +38,67 @@ export function useGameBacklogState({
     setIsInBacklogMaybeOptimistic(isInitiallyInBacklog);
   }, [isInitiallyInBacklog]);
 
-  const toggleBacklog = useCallback(async () => {
-    if (!auth?.user && typeof window !== 'undefined') {
-      window.location.href = route('login');
+  const toggleBacklog = useCallback(
+    async (options?: { shouldHideToasts: boolean }) => {
+      if (!auth?.user && typeof window !== 'undefined') {
+        window.location.href = route('login');
 
-      return;
-    }
-
-    const newBacklogState = !isInBacklogMaybeOptimistic;
-
-    // Only update state optimistically if configured to do so
-    if (shouldUpdateOptimistically) {
-      setIsInBacklogMaybeOptimistic(newBacklogState);
-    }
-
-    const mutationOptions: Parameters<typeof removeFromWantToPlayGamesList>[2] = {
-      shouldEnableToast: shouldShowToasts,
-    };
-
-    if (shouldShowToasts) {
-      // Add the success message when toasts are enabled.
-      mutationOptions.t_successMessage = newBacklogState
-        ? t('Added :gameTitle to playlist!', { gameTitle: game.title })
-        : t('Removed :gameTitle from playlist!', { gameTitle: game.title });
-
-      // Add the undo callback only when removing from backlog and toasts are enabled.
-      if (!newBacklogState) {
-        mutationOptions.onUndo = () => setIsInBacklogMaybeOptimistic(true);
-      }
-    }
-
-    try {
-      if (newBacklogState) {
-        await addToWantToPlayGamesList(game.id, game.title, mutationOptions);
-      } else {
-        await removeFromWantToPlayGamesList(game.id, game.title, mutationOptions);
+        return;
       }
 
-      // If we aren't updating optimistically, then update after successful mutation.
-      if (!shouldUpdateOptimistically) {
+      const shouldShowToasts = options?.shouldHideToasts !== true;
+      const newBacklogState = !isInBacklogMaybeOptimistic;
+
+      // Only update state optimistically if configured to do so
+      if (shouldUpdateOptimistically) {
         setIsInBacklogMaybeOptimistic(newBacklogState);
       }
-    } catch {
-      // We only need to revert if we're configured to update optimistically.
-      if (shouldUpdateOptimistically) {
-        setIsInBacklogMaybeOptimistic(!newBacklogState);
+
+      const mutationOptions: Parameters<typeof removeFromWantToPlayGamesList>[2] = {
+        shouldEnableToast: shouldShowToasts,
+      };
+
+      if (shouldShowToasts) {
+        // Add the success message when toasts are enabled.
+        mutationOptions.t_successMessage = newBacklogState
+          ? t('Added :gameTitle to playlist!', { gameTitle: game.title })
+          : t('Removed :gameTitle from playlist!', { gameTitle: game.title });
+
+        // Add the undo callback only when removing from backlog and toasts are enabled.
+        if (!newBacklogState) {
+          mutationOptions.onUndo = () => setIsInBacklogMaybeOptimistic(true);
+        }
       }
-    }
-  }, [
-    addToWantToPlayGamesList,
-    auth?.user,
-    game.id,
-    game.title,
-    isInBacklogMaybeOptimistic,
-    removeFromWantToPlayGamesList,
-    shouldShowToasts,
-    shouldUpdateOptimistically,
-    t,
-  ]);
+
+      try {
+        if (newBacklogState) {
+          await addToWantToPlayGamesList(game.id, game.title, mutationOptions);
+        } else {
+          await removeFromWantToPlayGamesList(game.id, game.title, mutationOptions);
+        }
+
+        // If we aren't updating optimistically, then update after successful mutation.
+        if (!shouldUpdateOptimistically) {
+          setIsInBacklogMaybeOptimistic(newBacklogState);
+        }
+      } catch {
+        // We only need to revert if we're configured to update optimistically.
+        if (shouldUpdateOptimistically) {
+          setIsInBacklogMaybeOptimistic(!newBacklogState);
+        }
+      }
+    },
+    [
+      addToWantToPlayGamesList,
+      auth?.user,
+      game.id,
+      game.title,
+      isInBacklogMaybeOptimistic,
+      removeFromWantToPlayGamesList,
+      shouldUpdateOptimistically,
+      t,
+    ],
+  );
 
   return { isInBacklogMaybeOptimistic, toggleBacklog, isPending };
 }
